LayoutTests/media/track/webvtt-ruby-expected.txt


EVENT(canplay)
EVENT(addtrack)
EXPECTED (video.textTracks.length == '1') OK
RUN(video.textTracks[0].mode = 'showing')
RUN(video.currentTime = 1)
EVENT(seeked)
EXPECTED (window.internals.shadowRoot(video).querySelector('rt') != 'null') OK
EXPECTED (rubyText.offsetTop < rubyBase.offsetTop == 'true') OK
EXPECTED (rubyBase.offsetTop == (rubyText.offsetTop + rubyText.offsetHeight) == 'true') OK
END OF TEST
